[editor] lift menu builder into shared EditorMenuService
The File/Edit/View/Help menu model and its commands move into a root
EditorMenuService, so the desktop title bar and the upcoming mobile menu
Drawer render the same items from one source. The title bar becomes a thin
consumer of the service's items signal.
